Реализация вращения, перемещения всем фигурам. Наследование от Figure

+++ src/Movable.kt
@@ -1,4 +1,4 @@
 interface Movable {
     // заказ на реализацию некоторых функций
-    fun move(dx: Int, dy: Int)
+    fun move(dx: Float, dy: Float)
 }
